retirei o long

+++ UFT/Estrutura de Dados/Playlist Youtube ED1/Exercicios Teste/Pilhas/PilhasEstaticas/stackestatic.c
@@ -4,8 +4,8 @@
 typedef struct _stackstatic{
 
     int top;
-    long capacity;
-    long *data;
+    int capacity;
+    int *data;
